+++ continuousintegration/docker/ctafrontend/cc7/ci_runner/prepareImage.sh
@@ -1,5 +1,7 @@
+#!/bin/bash -e
+
 # @project      The CERN Tape Archive (CTA)
-# @copyright    Copyright © 2015-2022 CERN
+# @copyright    Copyright © 2022 CERN
 # @license      This program is free software, distributed under the terms of the GNU General Public
 #               Licence version 3 (GPL Version 3), copied verbatim in the file "COPYING". You can
 #               redistribute it and/or modify it under the terms of the GPL Version 3, or (at your
@@ -13,15 +15,26 @@
 #               granted to it by virtue of its status as an Intergovernmental Organization or
 #               submit itself to any jurisdiction.
 
-# This module will set the following variables:
-#     OPENSSL_FOUND
-#     OPENSSL_INCLUDE_DIRS
+cd ~/CTA
 
-find_path (OPENSSL_INCLUDE_DIRS
-  openssl/rsa.h
-  PATHS /usr/include
-  NO_DEFAULT_PATH)
+rpm_source=$1
+rpm_dir="build_rpm/RPM/RPMS/x86_64"
 
-include(FindPackageHandleStandardArgs)
-find_package_handle_standard_args(openssl DEFAULT_MSG
-  OPENSSL_INCLUDE_DIRS)
+image_tag=$2
+
+if [ -z "${rpm_source}" ]; then
+  echo "You should specify the path to the RPMs to be installed. Ex: ~/CTA-build/RPM/RPMS/x86_64";
+  exit 1;
+fi
+
+if [ -z "${image_tag}" ]; then
+  echo "You should specify the docker image tag. Ex: dev";
+  exit 1;
+fi
+
+trap "rm -rf $rpm_dir" EXIT
+
+mkdir -p $rpm_dir
+cp -r $rpm_source $rpm_dir
+
+sudo docker build . -f continuousintegration/docker/ctafrontend/cc7/ci_runner/Dockerfile -t ctageneric:dev